separating models into apps
I separated models creating similar chunks folder to documents folder and connect them using routers in api.api file
