--- api/MyModels/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from .api import api
-
-urlpatterns = [
-    path("", api.urls),
-]
