+++ api/api/settings.py
@@ -29,7 +29,9 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
-    "MyModels.apps.DocumentsConfig",
+    "chunks.apps.ChunksConfig",
+    "documents.apps.DocumentsConfig",
+    
 ]
 
 MIDDLEWARE = [
