+++ api/api/urls.py
@@ -21,5 +21,5 @@ from django.conf.urls.static import static
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("api/", include("MyModels.urls")),
+    path("", include("api.api")),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
