+++ api/chunks/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from .models import Chunk
+
+admin.site.register(Chunk)
