+++ api/chunks/schemas.py
@@ -14,3 +14,4 @@ class ChunkOut(Schema):
     chunk_idx: int
     start_char: int
     end_char: int
+
